refactor: delete current_user_service and changed userProvider
