+++ lib/presentation/sign_up/sign_up_view_model.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:kovel_app/core/auth/current_user_service.dart';
 import 'package:kovel_app/domain/use_case/update_user_name_use_case.dart';
+
+import '../../core/auth/user_provider.dart';
 
 class SignUpViewModel with ChangeNotifier {
   final UpdateUserNameUseCase _updateUserNameUseCase;
